cryptography/SMS/app/requirements.txt
fastapi
uvicorn
pycrypto
